# api.yaml
openapi: 3.0.2
info:
  title: Pyrra
  version: 0.0.0
servers:
  - url: /api/v1
paths:
  /objectives:
    get:
      summary: List Objectives
      operationId: ListObjectives
      tags:
        - objectives
      responses:
        '200':
          description: List Objectives
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Objectives"
  /objectives/{namespace}/{name}:
    get:
      summary: Get Objective
      operationId: GetObjective
      tags:
        - objectives
      parameters:
        - in: path
          name: namespace
          required: true
          schema:
            type: string
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Get Objective
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Objective'
  /objectives/{namespace}/{name}/status:
    get:
      summary: Get objective status
      operationId: GetObjectiveStatus
      tags:
        - objectives
      parameters:
        - in: path
          name: namespace
          required: true
          schema:
            type: string
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Get objective status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectiveStatus'
  /objectives/{namespace}/{name}/errorbudget:
    get:
      summary: Get ErrorBudget graph sample pairs
      operationId: GetObjectiveErrorBudget
      tags:
        - objectives
      parameters:
        - in: path
          name: namespace
          required: true
          schema:
            type: string
        - in: path
          name: name
          required: true
          schema:
            type: string
        - in: query
          name: start
          schema:
            type: integer
          description: The timestamp to start the query range at
        - in: query
          name: end
          schema:
            type: integer
          description: The timestamp to end the query range at
      responses:
        '200':
          description: Get ErrorBudget graph sample pairs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryRange'
  /objectives/{namespace}/{name}/alerts:
    get:
      summary: Get the MultiBurnrateAlerts for the Objective
      operationId: GetMultiBurnrateAlerts
      tags:
        - objectives
      parameters:
        - in: path
          name: namespace
          required: true
          schema:
            type: string
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Get GetMultiBurnrateAlerts for the Objective
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MultiBurnrateAlert'
  /objectives/{namespace}/{name}/red/requests:
    get:
      summary: Get a matrix of requests by label
      operationId: GetREDRequests
      tags:
        - objectives
      parameters:
        - in: path
          name: namespace
          required: true
          schema:
            type: string
        - in: path
          name: name
          required: true
          schema:
            type: string
        - in: query
          name: start
          schema:
            type: integer
          description: The timestamp to start the query range at
        - in: query
          name: end
          schema:
            type: integer
          description: The timestamp to end the query range at
      responses:
        '200':
          description: Get a matrix of requests by label
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryRange'
  /objectives/{namespace}/{name}/red/errors:
    get:
      summary: Get a matrix of error percentage by label
      operationId: GetREDErrors
      tags:
        - objectives
      parameters:
        - in: path
          name: namespace
          required: true
          schema:
            type: string
        - in: path
          name: name
          required: true
          schema:
            type: string
        - in: query
          name: start
          schema:
            type: integer
          description: The timestamp to start the query range at
        - in: query
          name: end
          schema:
            type: integer
          description: The timestamp to end the query range at
      responses:
        '200':
          description: Get a matrix of error percentage by label
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryRange'
components:
  schemas:
    Objectives:
      type: array
      items:
        $ref: '#/components/schemas/Objective'
    Objective:
      type: object
      required:
        - name
        - namespace
        - description
        - target
        - window
        - config
      properties:
        name:
          type: string
        namespace:
          type: string
        description:
          type: string
        target:
          type: number
          format: double
        window:
          type: integer
          format: int64
        config:
          type: string
        indicator:
          $ref: '#/components/schemas/Indicator'
    Indicator:
      type: object
      properties:
        ratio:
          type: object
          required:
            - errors
            - total
          properties:
            errors:
              $ref: '#/components/schemas/Query'
            total:
              $ref: '#/components/schemas/Query'
        latency:
          type: object
          required:
            - success
            - total
          properties:
            success:
              $ref: '#/components/schemas/Query'
            total:
              $ref: '#/components/schemas/Query'
    Query:
      type: object
      required:
        - metric
      properties:
        metric:
          type: string
        name:
          type: string
        matchers:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              value:
                type: string
              type:
                type: integer
    ObjectiveStatus:
      type: object
      required:
        - availability
        - budget
      properties:
        availability:
          type: object
          required:
            - percentage
            - total
            - errors
          properties:
            percentage:
              type: number
              format: double
            total:
              type: number
              format: double
            errors:
              type: number
              format: double
        budget:
          type: object
          required:
            - total
            - remaining
            - max
          properties:
            total:
              type: number
              format: double
            remaining:
              type: number
              format: double
            max:
              type: number
              format: double
    MultiBurnrateAlert:
      type: object
      required:
        - severity
        - for
        - factor
        - short
        - long
        - state
      properties:
        severity:
          type: string
        for:
          type: integer
          format: int64
        factor:
          type: number
          format: double
        short:
          $ref: '#/components/schemas/Burnrate'
        long:
          $ref: '#/components/schemas/Burnrate'
        state:
          type: string
          enum:
            - inactive
            - pending
            - firing
    Burnrate:
      type: object
      required:
        - window
        - current
        - query
      properties:
        window:
          type: integer
          format: int64
        current:
          type: number
          format: double
        query:
          type: string
    QueryRange:
      type: object
      required:
        - query
        - labels
        - values
      properties:
        query:
          type: string
        labels:
          type: array
          items:
            type: string
        values:
          type: array
          items:
            type: array
            items:
              type: number
              format: double