# openapi/server/go/model_burnrate.go
/*
 * Pyrra
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 0.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type Burnrate struct {
	Window int64 `json:"window"`

	Current float64 `json:"current"`

	Query string `json:"query"`
}