# openapi/server/go/model_indicator.go
/*
 * Pyrra
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 0.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type Indicator struct {
	Ratio IndicatorRatio `json:"ratio,omitempty"`

	Latency IndicatorLatency `json:"latency,omitempty"`
}